requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.4.0
tqdm>=4.66.0
pyyaml>=6.0.1
opencv-python>=4.9.0
pillow>=10.0.0
streamlit>=1.28.0
pandas>=2.0.0